Module refctoring for plyers
