+++ src/game/tools.py
@@ -3,21 +3,6 @@ PIECES_NUMBER = 16
 EMPTY_POSITION = '.'
 
 
-class Players:
-
-    """Definition of the users names:"""
-
-    def __init__(self, initial_state=""):
-        self.player1_name = "Player 1"
-        self.player2_name = "Player 2"
-
-    def change_player_name(self, player_id, player_name):
-        if player_id == 1:
-            self.player1_name = player_name
-        if player_id == 2:
-            self.player2_name = player_name
-
-
 def get_coordinates(position):
     """Convert a postion of format 'A3' into coordinates x= 0 and y = 2 in the grid"""
     if len(position) != 2:
